Refactored sidebar into component
